% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runpkg.R
\name{download_main}
\alias{download_main}
\alias{download_files}
\title{Download files from npm packages via unpkg}
\usage{
download_main(pkg)

download_files(pkg, files = NULL)
}
\arguments{
\item{pkg}{the name (and optionally the version) of an npm package. (e.g. 'jquery', 'jquery@3.0.0')}

\item{files}{character vector of paths to files. These are used as the \code{:file}
part of \url{https://unpkg.com/:package@:version/:file}}
}
\value{
a \link[htmltools:htmlDependency]{htmltools::htmlDependency} object
}
\description{
Download files from npm packages via unpkg
}
\examples{

(fa <- download_main("fontawesome"))
htmltools::renderDependencies(list(fa), "href")
htmltools::renderDependencies(list(fa), "file")


(jquery <- download_files("jquery@3.0.0", "dist", "jquery.slim.min.js"))

}
\references{
\url{https://unpkg.com}
}
