% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runpkg.R
\name{ls_}
\alias{ls_}
\title{List package files}
\usage{
ls_(pkg, path = "/")
}
\arguments{
\item{pkg}{the name (and optionally the version) of an npm package. (e.g. 'jquery', 'jquery@3.0.0')}

\item{path}{character string with a path pointing to a package folder. This
is used as the \code{:file} part of \code{https://unpkg.com/:package@:version/:file}}
}
\value{
a \link[htmltools:htmlDependency]{htmltools::htmlDependency} object
}
\description{
List package files
}
\examples{

ls_("jquery")
ls_("jquery", "dist")
ls_("jquery@3.2.1", "external/sizzle")

}
\seealso{
\link{download_files}
}
