% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runpkg.R
\name{download_main}
\alias{download_main}
\alias{download_files}
\title{Download the 'main' file from a package}
\usage{
download_main(pkg)

download_files(pkg, files = NULL)
}
\arguments{
\item{pkg}{the name (and optionally the version) of an npm package. (e.g. 'jquery', 'jquery@3.0.0')}

\item{files}{character vector of paths to files (these paths should be relative to .}
}
\value{
a \link[htmltools:htmlDependency]{htmltools::htmlDependency} object

a \link[htmltools:htmlDependency]{htmltools::htmlDependency} object
}
\description{
Download the 'main' file from a package

Download npm package files and convert them to HTML dependencies
}
\examples{

(fa <- download_main("fontawesome"))
htmltools::renderDependencies(list(fa), "href")
htmltools::renderDependencies(list(fa), "file")


(jquery <- download_files("jquery@3.0.0", "dist", "jquery.slim.min.js"))

}
\references{
\url{https://unpkg.com/#/}
}
