% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runpkg.R
\name{ls_}
\alias{ls_}
\title{List package files}
\usage{
ls_(pkg, path = "/")
}
\arguments{
\item{pkg}{the name (and optionally the version) of an npm package. (e.g. 'jquery', 'jquery@3.0.0')}

\item{path}{a path to a folder.}
}
\value{
a \link[htmltools:htmlDependency]{htmltools::htmlDependency} object
}
\description{
List package files
}
\examples{

ls_("jquery")
ls_("jquery", "dist")
ls_("jquery@3.2.1", "external/sizzle")

}
\references{
\url{https://unpkg.com/#/}
}
